+++ packages/flutter_web_bluetooth_backend/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.1.0
+
+* Using newer ble_backend
+
 ## 1.0.0
 
 * Initial release