+++ packages/universal_ble_backend/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.0
+
+* Using newer ble_backend
+
 ## 0.1.0
 
 * Initial release